+++ src/components/app/App.module.css
@@ -1,0 +1,7 @@
+.content
+{ 
+    position: relative;
+    display: flex;
+    gap: 40px;
+    justify-content: center 
+}
