+++ src/components/burgerConstructor/BurgerConstr.js
@@ -4,102 +4,86 @@ import bgStyle from './BurgerConstructor.module.css'
 import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import img from '../../images/done.png'
+import productPropTypes from "../../utils/types";
 
-const productPropTypes = PropTypes.shape({
-  _id: PropTypes.string.isRequired,
-  __v: PropTypes.number,
+const Product = ({ data, type, isLocked, flag, name }) => {
+  return (
+    <li className={bgStyle.list}>
+      {flag && <DragIcon type="primary" />}
+      <ConstructorElement
+        type={type}
+        isLocked={isLocked}
+        text={name ? name : data.name}
+        price={data.price}
+        thumbnail={data.image}
+      >
+      </ConstructorElement>
+    </li>
+  );
+};
+
+Product.propTypes = {
+  data: productPropTypes,
   type: PropTypes.string,
-  name: PropTypes.string,
-  price: PropTypes.number,
-  proteins: PropTypes.number,
-  fat: PropTypes.number,
-  carbohydrates: PropTypes.number,
-  calories: PropTypes.number,
-  image: PropTypes.string,
-  image_large: PropTypes.string,
-  image_mobile: PropTypes.string
-});
+  isLocked: PropTypes.bool,
+  flag: PropTypes.bool,
+  name: PropTypes.string
+}
 
-const Product = ({data, type, isLocked, flag, name}) => {
-   return (
-      <li className={bgStyle.list}>
-        {flag&&<DragIcon type="primary" />}
-        <ConstructorElement
-            type={type}
-            isLocked={isLocked}
-            text={name?name:data.name}
-            price={data.price}
-            thumbnail={data.image}
-          >
-  </ConstructorElement>
-      </li>
-    );
-  };
-
-  Product.propTypes=PropTypes.shape({
-    data:productPropTypes,
-    type: PropTypes.string,
-    isLocked: PropTypes.bool,
-    flag:PropTypes.bool,
-    name:PropTypes.string
-  })
-
-class BurgerConstructor extends React.Component{
-  constructor(props){
+class BurgerConstructor extends React.Component {
+  constructor(props) {
     super(props);
-    this.data=props.data?props.data:[];
-    this.list=props.data?props.data:[];
-    this.price=this.list.map(i=>this.price+=i.price, this.price=0).reverse()[0];
-    
+    this.data = props.data ? props.data : [];
+    this.list = props.data ? props.data : [];
+    this.price = this.list.map(i => this.price += i.price, this.price = 0).reverse()[0];
+
     this.state = { show: false };
   }
 
-  showChanger =()=>{
+  showChanger = () => {
     console.log(this.state.show)
     this.setState({ show: !this.state.show })
   }
-    render()
-    {
-        return (
-          <div  className={bgStyle.main}>
-            
-           <ul className={`custom-scroll`}>
-           {this.list.map((prod, index)=>{
-              if(prod.type=="bun" ){
-                if (index==0){
-            return <Product key={index} type={"top"} data={prod} isLocked={true} flag={false} name="Краторная булка N-200i (верх)"/>}}})}
-           <li>
-           {this.list.map((prod, index)=>{
-              if(prod.type=="sauce" ){
-                 return <Product key={index} data={prod} isLocked={false} flag={true}/>   
+  render() {
+    return (
+      <div className={bgStyle.main}>
+
+        <ul className={`custom-scroll`}>
+          {this.list.map((prod) => {
+            if (prod._id == "60666c42cc7b410027a1a9b1") {
+              return <Product key={prod._id} type={"top"} data={prod} isLocked={true} flag={false} name="Краторная булка N-200i (верх)" />
+            }
+          })}
+            {this.list.map((prod) => {
+              if (prod.type == "sauce") {
+                return <Product key={prod._id} data={prod} isLocked={false} flag={true} />
               }
-            }  )}
-           </li>
-           {this.list.map((prod, index)=>{
-              if(prod.type=="bun" ){
-                if (index==0){
-            return <Product key={index} type={"bottom"} data={prod} isLocked={true} flag={false} name="Краторная булка N-200i (низ)"/>}}})}
-          </ul>
-          <div className={bgStyle.bottom}><div className={bgStyle.price}>{this.price} <CurrencyIcon/></div>
+            })}
+          {this.list.map((prod, index) => {
+            if (prod._id == "60666c42cc7b410027a1a9b1")
+              return <Product key={prod._id} type={"bottom"} data={prod} isLocked={true} flag={false} name="Краторная булка N-200i (низ)" />
+          })}
+        </ul>
+        <div className={bgStyle.bottom}><div className={bgStyle.price}>{this.price} <CurrencyIcon /></div>
           <Button type="primary" size="medium" className={bgStyle.but} onClick={this.showChanger}>
-          Оформить заказ
-        </Button>
-        
-        {this.state.show&&
-          <div onClick={this.showChanger}>
-          <Modal >
-          <Child/>
-        </Modal>
-        </div>}
+            Оформить заказ
+          </Button>
+
+          {this.state.show &&
+            <div onClick={this.showChanger}>
+              <Modal >
+                <Child />
+              </Modal>
+            </div>}
         </div>
-          
-        </div>
-        )
-    }
+
+      </div>
+    )
+  }
 }
 
-BurgerConstructor.propTypes={
-  data:PropTypes.arrayOf(productPropTypes)
+BurgerConstructor.propTypes = {
+  data: PropTypes.arrayOf(productPropTypes)
 }
 
 export default BurgerConstructor;
@@ -112,7 +96,7 @@ class Modal extends React.Component {
     super(props);
     this.el = document.createElement('div');
   }
-  
+
   componentDidMount() {
     // Элемент портала добавляется в DOM-дерево после того, как
     // потомки компонента Modal будут смонтированы, это значит,
@@ -137,8 +121,8 @@ class Modal extends React.Component {
   }
 }
 
-Modal.propTypes={
-  data:productPropTypes
+Modal.propTypes = {
+  data: productPropTypes
 }
 
 function Child() {
@@ -147,16 +131,16 @@ function Child() {
   return (
     <section className={bgStyle.overlay} >
       <div className={bgStyle.details}>
-      <Button type="secondary" size="small" onClick={() => console.log("hi")}>Х</Button>
-      <p className="text text_type_digits-large" style={{textShadow: " 0px 0px 16px rgba(51, 51, 255, 0.25), 0px 0px 8px rgba(51, 51, 255, 0.25), 0px 4px 32px rgba(51, 51, 255, 0.5)"}}>123456</p>
-      <h1>Идентификатор заказа</h1>
-      <img src={img} alt="error"/>
-      <p>Ваш заказ начали готовить</p>
-      <p style={{color: "#8585AD"}}>Дождитесь готовности на орбитальной станции</p>
-    </div>
+        <Button type="secondary" size="small" onClick={() => console.log("hi")}>Х</Button>
+        <p className={`text text_type_digits-large ${bgStyle.num}`}>123456</p>
+        <h1>Идентификатор заказа</h1>
+        <img src={img} alt="error" />
+        <p>Ваш заказ начали готовить</p>
+        <p className={bgStyle.disabled}>Дождитесь готовности на орбитальной станции</p>
+      </div>
     </section>
   );
 }
-Child.propTypes={
-  data:productPropTypes
+Child.propTypes = {
+  data: productPropTypes
 }
