+++ src/components/burgerConstructor/BurgerConstructor.module.css
@@ -73,4 +73,10 @@
     align-items: center;
     border-radius: 40px;
   }
-  
+  .disabled{
+    
+  color: #8585AD;
+  }
+  .num{
+    text-shadow: 0px 0px 16px rgba(51, 51, 255, 0.25), 0px 0px 8px rgba(51, 51, 255, 0.25), 0px 4px 32px rgba(51, 51, 255, 0.5);
+  }
