+++ src/components/burgerIngredients/BurgerStyle.module.css
@@ -179,3 +179,9 @@
     width: 80%;
   }
 }
+.child{
+  display:flex;
+  flex-direction:row;
+  justify-content:space-between;
+  width:640px;
+}
